pyproject.toml
[project]
name = "docqa-system"
version = "0.1.0"
description = "An intelligent document Q&A system"
authors = [{name = "Developer", email = "developer@example.com"}]
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    "streamlit>=1.28.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.15",
    "langchain-openai>=0.0.5",
    "langchain-chroma>=0.1.0",
    "pypdf>=3.17.0",
    "tiktoken>=0.5.1",
    "python-dotenv>=1.0.0",
    "chardet>=5.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]